chore: refactor structure for api version
