+++ apps/v1/refresh_tokens/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class AuthenticationConfig(AppConfig):
+class RefreshTokensConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "apps.authentication"
+    name = "apps.v1.refresh_tokens"
