+++ apps/v1/refresh_tokens/models.py
@@ -1,7 +1,7 @@
 from django.db import models
 from django.utils.translation import gettext_lazy as _
 
-from apps.authentication.models import User
+from apps.v1.authentication.models import User
 
 
 class RefreshTokenFamilyStatus(models.TextChoices):
