+++ apps/v1/urls.py
@@ -1,4 +1,4 @@
-"""project_name URL Configuration
+"""project_name API v1 URL Configuration
 
 The `urlpatterns` list routes URLs to views. For more information please see:
     https://docs.djangoproject.com/en/3.2/topics/http/urls/
@@ -14,15 +14,19 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 
-from django.contrib import admin
-from django.http import HttpResponse
 from django.urls import include, path
+from drf_spectacular.views import (
+    SpectacularAPIView,
+    SpectacularRedocView,
+    SpectacularSwaggerView,
+)
 
 urlpatterns = [
-    # health
-    path("api/health", lambda request: HttpResponse("OK")),
-    # admin
-    path("admin/", admin.site.urls),
+    # docs
+    path("schema/", SpectacularAPIView.as_view(), name="schema"),
+    path("docs/", SpectacularSwaggerView.as_view(url_name="schema"), name="swagger-ui"),
+    path("redoc/", SpectacularRedocView.as_view(url_name="schema"), name="redoc"),
     # apis
-    path("api/v1/", include("apps.v1.urls")),
+    path("auth/", include("apps.v1.authentication.urls")),
+    path("auth/refresh-token", include("apps.v1.refresh_tokens.urls")),
 ]
